In INDEX, try FILE-POSITION for all types of streams
The previous dispatch type, FILE-STREAM, excluded things like
FLEXI-STREAM:IN-MEMORY-*-STREAM which actually provide a working
FILE-POSITION.

Calling FILE-POSITION on any stream is no worse than the previous
behavior which returned 0 and could produce corrupt archives when
encoding. For the new code, if the stream does not support
FILE-POSITION and returns an arbitrary value, the result may be a
corrupt archive as before. If the stream does not support
FILE-POSITION and an error is signaled, silent corruption is avoided,
which seems better.